+++ gatsby-ssr.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export const onRenderBody = ({ setPreBodyComponents }) => {
+  setPreBodyComponents([<html-comment></html-comment>])
+}