+++ content/pages/changelog.md
@@ -4,6 +4,7 @@ slug: changelog
 template: page
 ---
 
+- 29-07-22: Added html comment Easter egg
 - 28-07-22: Added fuzzy mode
 - 26-07-22: Added simple changelog page
 - 25-07-22: Added new footnotes and marginalia formatting to markdown posts